[hotfix][runtime-tests] Deduplicate CollectingResultPartitionWriters classes
